deploy 1 ready

+++ apps/ws-backend/src/events/onElementDelete.ts
@@ -2,6 +2,7 @@ import { room } from "../RoomManager";
 import { ClientMessage } from "@repo/common/messageTypes";
 import { prismaClient } from "@repo/db/client";
 import WebSocket from 'ws';
+import { checkAccess } from "../utils/checkAccess";
 
 export async function onElementDelete(ws: WebSocket, message: ClientMessage) {
     if (message.type !== 'ELEMENT_DELETE') return;
@@ -11,6 +12,14 @@ export async function onElementDelete(ws: WebSocket, message: ClientMessage) {
     const senderId = room.getUserId(ws);
     
     if (!senderId) return;
+
+    if (!checkAccess(senderId, roomId)) {
+        ws.send(JSON.stringify({
+            type: 'ERROR',
+            message: 'Unauthorized Accesss'
+        }));
+        return;
+    }
     
     await prismaClient.element.deleteMany({
         where: {
